pyproject.toml
[tool.poetry]
name = "travel-anp-agent"
version = "0.1.0"
description = "Travel ANP Agent"
authors = ["Author <author@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
agent-connect = "^0.3.5"
pyjwt = "^2.10.1"
pydantic-settings = "^2.8.0"
aiohttp = "^3.11.13"
fastapi = "^0.112.1"
uvicorn = "^0.32.0"
ecdsa = "^0.19.0"
cryptography = "^43.0.3"
asn1crypto = "^1.5.1"
base58 = "^2.1.1"
websockets = "^13.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
